fix(master): correccion decimales

+++ service/orderService.go
@@ -22,11 +22,11 @@ func (p PairList) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
 func (p PairList) Less(i, j int) bool { return p[i].Value < p[j].Value }
 
 // orderMap : function to sort map by value
-func orderMap(noble *map[string]int) PairList {
+func orderMap(noble *map[string]float64) PairList {
 	p := make(PairList, len(*noble))
 	i := 0
 	for k, v := range *noble {
-		p[i] = Pair{k, v}
+		p[i] = Pair{k, int(v)}
 		i++
 	}
 	sort.Sort(p)
